services/fraud-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
azure-servicebus==7.12.1
opencensus-ext-azure==1.1.13
opencensus-ext-fastapi==0.7.2
python-dotenv==1.0.1